requirements.txt
anyio==3.5.0
asgiref==3.5.0
atomicwrites==1.4.0
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
connection==2021.7.20
coverage==6.3.2
dnspython==2.2.1
email-validator==1.1.3
fastapi==0.75.0
googlemaps==4.6.0
h11==0.13.0
httptools==0.3.0
idna==3.3
iniconfig==1.1.1
neo4j==4.4.1
numpy==1.22.3
packaging==21.3
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pydantic==1.9.0
pyparsing==3.0.7
pytest==7.1.1
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
sniffio==1.2.0
starlette==0.17.1
tomli==2.0.1
typing-extensions==4.1.1
urllib3==1.26.9
uvicorn==0.17.5
watchgod==0.7
websockets==10.2
